Refactor host-specific implementations and add missing functionality
- Implemented host-specific functions for unported game features in `unported.c`, including handling of game states and memory leaks.
- Enhanced `state_recorder.cpp` to support frame sinks for display copies, allowing for better frame management.
- Updated `main.cpp` to report title scene statistics and handle title scene execution.
- Created `absent_devices.c` to simulate the absence of console peripherals like memory cards and USB adapters.
- Improved alarm handling in `alarm.c` with a new function to check for the next alarm.
- Added support for freezing and advancing the OS clock in `dolphin_time.cpp`.
- Modified `video.cpp` to ensure proper handling of retraces and draw-done fences.
- Added tests for the new functionality in `gx_state_test.cpp`, `os_clock_test.cpp`, and `video_test.cpp`.
- Adjusted various files in the melee source to accommodate host-specific behavior, including stack reporting and alarm handling.
- Ensured proper handling of sound table reading in `axdriver.c` for host environments.
- Updated floating-point parsing in `fobj.c` to avoid undefined behavior in host builds.
- Enhanced texture expression handling in `texp.c` to prevent null pointer dereferencing.

+++ extern/dolphin/src/dolphin/axfx/chorus.c
@@ -409,9 +409,9 @@ L_00000344:
 
 int AXFXChorusInit(struct AXFX_CHORUS* c)
 {
-    long* left;
-    long* right;
-    long* sur;
+    s32* left;
+    s32* right;
+    s32* sur;
     u32 i;
     int old;
 
@@ -479,12 +479,12 @@ int AXFXChorusSettings(struct AXFX_CHORUS* c)
 void AXFXChorusCallback(struct AXFX_BUFFERUPDATE* bufferUpdate,
                         struct AXFX_CHORUS* chorus)
 {
-    long* leftD;
-    long* rightD;
-    long* surD;
-    long* leftS;
-    long* rightS;
-    long* surS;
+    s32* leftD;
+    s32* rightD;
+    s32* surD;
+    s32* leftS;
+    s32* rightS;
+    s32* surS;
     u32 i;
     u8 nextCurrentLast;
